Replace AssemblerSpec with AssemblerTests.
